+++ packages/ui/src/components/Statements/StatementPage.tsx
@@ -2,6 +2,7 @@ import { __internal_useStatementQuery } from '@clerk/shared/react/index';
 
 import { Alert } from '@/ui/elements/Alert';
 import { Header } from '@/ui/elements/Header';
+import { ProfileCard } from '@/ui/elements/ProfileCard';
 import { formatDate } from '@/ui/utils/formatDate';
 
 import { useSubscriberTypeContext, useSubscriberTypeLocalizationRoot } from '../../contexts/components';
@@ -15,7 +16,7 @@ import {
   Spinner,
   useLocalizations,
 } from '../../customizables';
-import { ArrowRightIcon, Plus, RotateLeftRight } from '../../icons';
+import { ArrowRight, Plus, RotateLeftRight } from '../../icons';
 import { useRouter } from '../../router';
 import { Statement } from './Statement';
 
@@ -38,18 +39,20 @@ export const StatementPage = () => {
 
   if (isLoading) {
     return (
-      <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100%' }}>
-        <Spinner
-          colorScheme='primary'
-          sx={{ margin: 'auto', display: 'block' }}
-          elementDescriptor={descriptors.spinner}
-        />
-      </Box>
+      <ProfileCard.Page>
+        <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100%' }}>
+          <Spinner
+            colorScheme='primary'
+            sx={{ margin: 'auto', display: 'block' }}
+            elementDescriptor={descriptors.spinner}
+          />
+        </Box>
+      </ProfileCard.Page>
     );
   }
 
   return (
-    <>
+    <ProfileCard.Page>
       <Header.Root
         sx={t => ({
           borderBlockEndWidth: t.borderWidths.$normal,
@@ -130,7 +133,7 @@ export const StatementPage = () => {
                             >
                               <Span localizationKey={localizationKeys('billing.viewPayment')} />
                               <Icon
-                                icon={ArrowRightIcon}
+                                icon={ArrowRight}
                                 size='sm'
                                 aria-hidden
                               />
@@ -170,6 +173,6 @@ export const StatementPage = () => {
           />
         </Statement.Root>
       )}
-    </>
+    </ProfileCard.Page>
   );
 };